Added in UI and Datamanagment
Some things to note, I have created a scene testing ground for the UI in question so that it can be run, edited and tested, but for the sake of this commit I also created a prefab of the menu in the prefabs folder with a subfolder called UI that contains "VisualCanvas."

The intention for "VisualCanvas" is to store all kinds of UI elements such as the overlay in-game UI and menu UI as two examples.

Datamangement is WIP, current version does not create any compilation nor runtime errors, however it is advised not to mess with namespace as I am handling it as you read.